smaller geojson
